pyproject.toml
[project]
name = "claude-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.2.0",
    "fastmcp>=2.10.1",
    "google-genai>=1.24.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
